# levenshtein: wrap to sha256 for state hashing