--- a/go.mod
+++ b/go.mod
@@ -4,7 +4,7 @@ go 1.21
 
 require (
 	github.com/bits-and-blooms/bitset v1.2.0
-	github.com/blevesearch/mmap-go v1.0.4
+	github.com/edsrzf/mmap-go v1.2.1-0.20241212181136-fad1cd13edbd
 	github.com/spf13/cobra v1.4.0
 )