@CascadingRadium 2026-01-22
superseded by https://github.com/blevesearch/vellum/pull/25